fix(test): mock getClientRect
Also move mocks from setup.mjs into testHelpers.
